Change PDF form field font
Use iText to change the font of interactive form fields in a PDF.

Changes the font in 'f1040.pdf' to Courier.

This code is based on the following example from iText:
https://kb.itextsupport.com/home/it7kb/ebooks/itext-7-jump-start-tutorial-for-net/chapter-5-manipulating-an-existing-pdf-document-net#Chapter5:ManipulatinganexistingPDFdocument|.NET-Changingthepropertiesofformfields

For now I just downloaded the dependencies' JAR files and put them in a
'lib/' directory rather than have to learn how to use a dependency
management tool like Maven. Here is the contents of the 'lib/'
directory:

    barcodes-7.2.2.jar
    commons-7.2.2.jar
    font-asian-7.2.2.jar
    forms-7.2.2.jar
    hyph-7.2.2.jar
    io-7.2.2.jar
    kernel-7.2.2.jar
    layout-7.2.2.jar
    pdfa-7.2.2.jar
    pdftest-7.2.2.jar
    sign-7.2.2.jar
    slf4j-api-1.7.9.jar
    styled-xml-parser-7.2.2.jar
    svg-7.2.2.jar

To run this:

    $ make
    $ make run