Makefile: Move shell script wrapper template exec to a new target
When I tried packaging with MacPorts, the path written to the shell
wrapper was the temporary MacPorts build path, not the final install
path.

My plan is to extract the shell wrapper build to the `package` stage,
and do something with the variables to give it the correct install path.
Not sure what that something is yet, but that's the idea.